refactor: move amm math into its own crate
