+++ state-chain/amm-math/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "cf-amm-math"
+version = "0.1.0"
+authors = ["Chainflip <https://chainflip.io>"]
+edition = "2021"
+description = "Chainflip AMM Math Primitives"
+
+[dependencies]
+cf-primitives = { workspace = true }
+serde = { workspace = true, features = ["derive", "alloc"] }
+
+# Parity deps
+codec = { workspace = true, features = ["derive"] }
+scale-info = { workspace = true, features = ["derive"] }
+
+sp-core = { workspace = true }
+sp-std = { workspace = true }
+
+rand = { workspace = true, features = ["std"], optional = true }
+
+[features]
+default = ["std"]
+slow-tests = ["dep:rand"]
+std = [
+    "cf-primitives/std",
+    "codec/std",
+    "scale-info/std",
+    "sp-core/std",
+    "sp-std/std",
+    "serde/std",
+]
