+++ state-chain/pallets/cf-pools/src/lib.rs
@@ -1,7 +1,8 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 use cf_amm::{
-	common::{self, Amount, PoolPairsMap, Price, Side, SqrtPriceQ64F96, Tick},
+	common::{PoolPairsMap, Side},
 	limit_orders::{self, Collected, PositionInfo},
+	math::{bounded_sqrt_price, Amount, Price, SqrtPriceQ64F96, Tick},
 	range_orders::{self, Liquidity},
 	PoolState,
 };
@@ -159,8 +160,8 @@ pub const PALLET_VERSION: StorageVersion = StorageVersion::new(5);
 #[frame_support::pallet]
 pub mod pallet {
 	use cf_amm::{
-		common::Tick,
 		limit_orders,
+		math::Tick,
 		range_orders::{self, Liquidity},
 		NewError,
 	};
@@ -752,7 +753,7 @@ pub mod pallet {
 								side,
 								id,
 								previous_tick,
-								IncreaseOrDecrease::Decrease(cf_amm::common::Amount::MAX),
+								IncreaseOrDecrease::Decrease(Amount::MAX),
 								NoOpStatus::Error,
 							)?;
 							Self::inner_update_limit_order(
@@ -827,7 +828,7 @@ pub mod pallet {
 							side,
 							id,
 							previous_tick,
-							IncreaseOrDecrease::Decrease(cf_amm::common::Amount::MAX),
+							IncreaseOrDecrease::Decrease(Amount::MAX),
 							NoOpStatus::Error,
 						)?;
 
@@ -1024,9 +1025,7 @@ pub mod pallet {
 										side,
 										id,
 										previous_tick,
-										crate::pallet::IncreaseOrDecrease::Decrease(
-											cf_amm::common::Amount::MAX,
-										),
+										crate::pallet::IncreaseOrDecrease::Decrease(Amount::MAX),
 										crate::NoOpStatus::Allow,
 									)?;
 									Ok(())
@@ -1080,6 +1079,8 @@ impl<T: Config> SwappingApi for Pallet<T> {
 		to: any::Asset,
 		input_amount: AssetAmount,
 	) -> Result<AssetAmount, DispatchError> {
+		use cf_amm::math::tick_at_sqrt_price;
+
 		let (asset_pair, order) =
 			AssetPair::from_swap(from, to).ok_or(Error::<T>::PoolDoesNotExist)?;
 		Self::try_mutate_pool(asset_pair, |_asset_pair, pool| {
@@ -1101,13 +1102,12 @@ impl<T: Config> SwappingApi for Pallet<T> {
 					.ok_or(Error::<T>::InsufficientLiquidity)?
 					.2;
 
-				let swap_tick = common::tick_at_sqrt_price(
-					PoolState::<(T::AccountId, OrderId)>::swap_sqrt_price(
+				let swap_tick =
+					tick_at_sqrt_price(PoolState::<(T::AccountId, OrderId)>::swap_sqrt_price(
 						order,
 						input_amount,
 						output_amount,
-					),
-				);
+					));
 				let bounded_swap_tick = if tick_after < tick_before {
 					core::cmp::min(core::cmp::max(tick_after, swap_tick), tick_before)
 				} else {
@@ -1383,8 +1383,8 @@ impl<T: Config> Pallet<T> {
 		lp: &T::AccountId,
 		side: Side,
 		id: OrderId,
-		tick: cf_amm::common::Tick,
-		sold_amount: cf_amm::common::Amount,
+		tick: Tick,
+		sold_amount: Amount,
 		noop_status: NoOpStatus,
 	) -> Result<(Collected, PositionInfo), DispatchError> {
 		let (collected, position_info) = match pool.pool_state.collect_and_mint_limit_order(
@@ -1419,8 +1419,8 @@ impl<T: Config> Pallet<T> {
 		asset_pair: &AssetPair,
 		side: Side,
 		id: OrderId,
-		tick: cf_amm::common::Tick,
-		sold_amount_change: IncreaseOrDecrease<cf_amm::common::Amount>,
+		tick: Tick,
+		sold_amount_change: IncreaseOrDecrease<Amount>,
 		noop_status: NoOpStatus,
 	) -> Result<AssetAmount, DispatchError> {
 		let (sold_amount_change, position_info, collected) = match sold_amount_change {
@@ -1495,7 +1495,7 @@ impl<T: Config> Pallet<T> {
 		lp: &T::AccountId,
 		asset_pair: &AssetPair,
 		id: OrderId,
-		tick_range: Range<cf_amm::common::Tick>,
+		tick_range: Range<Tick>,
 		size_change: IncreaseOrDecrease<range_orders::Size>,
 		noop_status: NoOpStatus,
 	) -> Result<(AssetAmounts, Liquidity), DispatchError> {
@@ -1717,7 +1717,7 @@ impl<T: Config> Pallet<T> {
 	pub fn required_asset_ratio_for_range_order(
 		base_asset: any::Asset,
 		quote_asset: any::Asset,
-		tick_range: Range<cf_amm::common::Tick>,
+		tick_range: Range<Tick>,
 	) -> Result<PoolPairsMap<Amount>, DispatchError> {
 		let pool_state = Pools::<T>::get(AssetPair::try_new::<T>(base_asset, quote_asset)?)
 			.ok_or(Error::<T>::PoolDoesNotExist)?
@@ -1768,7 +1768,7 @@ impl<T: Config> Pallet<T> {
 						} else {
 							Some(PoolOrder {
 								amount: sold_base_amount,
-								sqrt_price: cf_amm::common::bounded_sqrt_price(
+								sqrt_price: bounded_sqrt_price(
 									bought_quote_amount,
 									sold_base_amount,
 								),
@@ -1794,7 +1794,7 @@ impl<T: Config> Pallet<T> {
 						} else {
 							Some(PoolOrder {
 								amount: bought_base_amount,
-								sqrt_price: cf_amm::common::bounded_sqrt_price(
+								sqrt_price: bounded_sqrt_price(
 									sold_quote_amount,
 									bought_base_amount,
 								),
@@ -1809,7 +1809,7 @@ impl<T: Config> Pallet<T> {
 	pub fn pool_depth(
 		base_asset: any::Asset,
 		quote_asset: any::Asset,
-		tick_range: Range<cf_amm::common::Tick>,
+		tick_range: Range<Tick>,
 	) -> Result<AskBidMap<UnidirectionalPoolDepth>, DispatchError> {
 		let asset_pair = AssetPair::try_new::<T>(base_asset, quote_asset)?;
 		let mut pool = Pools::<T>::get(asset_pair).ok_or(Error::<T>::PoolDoesNotExist)?;
