+++ state-chain/primitives/src/lib.rs
@@ -10,7 +10,7 @@ use frame_support::sp_runtime::{
 };
 use scale_info::TypeInfo;
 use serde::{Deserialize, Serialize};
-use sp_core::{ConstU32, U256};
+use sp_core::ConstU32;
 use sp_std::{
 	cmp::{Ord, PartialOrd},
 	fmt,
@@ -102,20 +102,6 @@ pub type PrewitnessedDepositId = u64;
 pub type BoostPoolTier = u16;
 
 define_wrapper_type!(AffiliateShortId, u8, extra_derives: Serialize, Deserialize, PartialOrd, Ord);
-
-// TODO: Consider alternative representation for Price:
-//
-// increasing Price to U512 or switch to a f64 (f64 would only be for the external
-// price representation), as at low ticks the precision in the price is VERY LOW, but this does not
-// cause any problems for the AMM code in terms of correctness
-
-/// This is the ratio of equivalently valued amounts of asset One and asset Zero.
-///
-/// The price is always measured in amount of asset One per unit of asset Zero. Therefore as asset
-/// zero becomes more valuable relative to asset one the price's literal value goes up, and vice
-/// versa. This ratio is represented as a fixed point number with `PRICE_FRACTIONAL_BITS` fractional
-/// bits.
-pub type Price = U256;
 
 /// The type of the Id given to threshold signature requests. Note a single request may
 /// result in multiple ceremonies, but only one ceremony should succeed.
